Add/Remove custom and inexistent plans in prod
We have some plans created by tenants, specifically the ones provided
by the re-observe team for prometheus.

These services generate events that we capture in our billing, and
we need the corresponding plan. In this commit we update the plan
with the reported unique_id as found in prod.

We found out the right list by querying:

 - Existing unique_id in prod (Run both ie and london):

   select distinct unique_id from service_plans
   where unique_id not like 'CAT%' order by 1;

 - unique_ids in this file:

   cat ./config.json  | jq -r '.pricing_plans[].plan_guid' |sort

Then we compared the output with diff

The plan for "redis tiny (compose)" that does not seem to be in
the `service_plans` in prod, which we might need to solve manually.