Map unknown services to a generic "Unknown Plan"
In production there are some events that are pointing to a plan_id
that is not recorded in the `services_plans` databases.

This prevents the computation of events to complete due the
restriction that `event.plan_unique_id` must not be null.

This situation might happen for events of resources created before
they were recorded in `services_plans`, or if there are events
for a resource that has not been fetched yet.

To fix it, we add a default plan unique_id that would be added
to any event without a corresponding `service_plan`.

The uuid of this new plan is d5091c33-2f9d-4b15-82dc-4ad69717fc03 and
we add it to config.json with a cost of 0.