Map unknown services to a generic "Unknown Plan"
There might be some events that are pointing to a
`service_plan_guid` that is not recorded in the `services_plans`
table, but `event.plan_guid` must not be null.

This situation might happen because:
 - events of resources created before we added retrieval of
   of the catalogs into `service_plans`
 - events for a resource which `service_plans` info has not
   been fetched yet.

For these cases, we add a default plan unique_id that would be added
to any `service` event with a `service_plan_guid`
without a corresponding `service_plan`.

The uuid of this new plan is d5091c33-2f9d-4b15-82dc-4ad69717fc03 and
we add it to config.json with a cost of 0.