IgnoreMissingPlans creates plans with unique_id
The flag `IgnoreMissingPlans` creates pricing_plans on the fly if
there is none defined already. This allows us to solve the situation
that a new broker is added but there is not a corresponding plan
for it. This is specially a problem for user provided brokers.

The new plans need to be created based on the plan_unique_id of
the events, rather than the plan_id.

We also add tests around the check of plans consistency and
this IgnoreMissingPlans feature.