IgnoreMissingPlans avoids creating duplicate plans
The flag `IgnoreMissingPlans` creates pricing_plans on the fly if
with start date to `epoch` (`1970-01-01`).
This allows us to solve the situation that a new broker is added
but there is not a corresponding plan for it, which is specially a
problem for user provided brokers.

But when we added a new "Unknown plan" to our config.json starting
on the `epoch`, this code failed due a duplicated entry:

    {
       "message" : "paas-billing.exit-error",
       "data" : {
          "error" : "generate-service-plan: duplicate key value violates unique constraint \"pricing_plans_pkey\": Key (plan_guid, valid_from)=(d5091c33-2f9d-4b15-82dc-4ad69717fc03, 1970-01-01 00:00:00+00) already exists."
       },
       "timestamp" : "1535650599.951504469",
       "source" : "paas-billing",
       "log_level" : 2
    }

To prevent this we could change the `valid_from` of the unknown plan
but instead we add a check in the generation of missing plans
to not generate a plan if there is already one in `epoch`.

We also extend the corresponding tests to cover the existing
feature and this specific case.