Set unique_id for orphan "redis tiny (compose)"
We found out that in prod (Ireland) there are some events referring
to a service plan_guid 53ca5c56-5474-4d64-9211-fe9aee86d502,
but there is no corresponding entry in the historic of services
in `services` and `service_plans` table.

With the new logic these events would be mapped to the "Unknown plan"
and have a cost of 0.

To workaround this, we add manually and adhoc a new entry in
`services` and `service_plans` for this service, with a new
uuid that we have pulled from the paas-cf history[1]:

    a8574a4b-9c6c-40ea-a0df-e9b7507948c8

The script in scripts/create_adhoc_redis_service.sql must be run
manually on prod billing-db, for example using cf conduit.

[1] https://github.com/alphagov/paas-cf/commit/b3f8018795c55f0f0bd36ed9e7b48d20bc76d87b