Add test helper to build complex scenarios
We'll use this helper in a future commit - the intent is to reduce
duplication and the need to compare hard-coded GUIDs. The scenario will
keep track of the GUIDs of each application, plan, event, space and org
by their names. It also provides a helper to quickly create a series of
events from an app's lifecycle.