Consolidate past billable events
Historical bills should not change if we change the current pricing
plan.  To ensure that tenants see the same bill in this situation we're
adding two new tables:

* consolidation_history - stores the date ranges which we've
  consolidated
* consolidated_billable_events - stores all the billable events (and
  their prices) (foreign key relation to consolidation_history)

We've added a `Consolidate` function which will be called periodically to
consolidate billable events across all organisations. We're also storing
the time that `Consolidate` ran in case we want to allow multiple
versions of consolidated billing events to co-exist (for example if we
change a historical pricing decision and need to record this).

There's a `IsRangeConsolidated` function which says whether a range
filter has been consolidated, and a `GetConsolidatedBillableEvents`
function which will return the events from the consolidated table.

`GetConsolidatedBillableEvents` should always return the same kind of
rows as the pre-existing `GetBillableEvents` function.