ConsolidateFullMonths and ConsolidateAll
ConsolidateFullMonths will over all full months (from 1st to 1st)
of the given range, truncating the date to the 1st of each month.

ConsolidateSince runs from of the date we started adding
records (2017-07-01) and the previous month when it is after
the 5th. The range can be changed sing the env vars
CONSOLIDATION_START_DATE and CONSOLIDATION_END_DATE.

For each month it will check whether there has already been a
consolidation, and if not run one.

Tests check that it runs for a given range, and that if it's run a
second time it will not overwrite the data it stored the first time
(even if the raw events or pricing plans have changed).

We add some functions to EventFilter to allow split and truncate it.