Remove plan_guid from events and use PlanUniqueID
Remove the plan_guid and (eventio.UsageEvent.PlanID), as
the value is not used anymore since we use
`plan_unique_id` and having both is confusing.

This has the consequence that the forecast events handler would
change behaviour, as it needs to pass an event with `plan_unique_id`.
We can add later a temporary implementation to support
the legacy format.