Service events include catalog as plan_guid
We want to identify and match the `service` events based on the unique
ID reported by the catalog, rather than the `service_plan_id` which
is specific per CF deployment.

The unique_id is periodically pulled from the CF api and stored
in the service_plans table.

Any other type of event (app, staging, tasks...)  would use the
plan_guid reported in the subquery above.

We also fix all the tests accordingly. We need to inject a new
entry in the `service_plans` table as needed.