Run ConsolidateAll in the EventProcessor and start
We want to run the consolidation periodically and ensure that we
run it after refresh all the events, and also when the application
starts.

We run it as part of the EventProcessor every time after the
events refresh(). The ConsolidateSince() will be a light operation
when it checks if each month is consolidated already.

On start, we run it after retrieving all the historic information.

In order to test this logic, we extract the main loop to a function
where we can inject a fake store to test.