Use context in read transactions
The methods:
 - GetBillableEventRows
 - GetConsolidatedBillableEventRows
 - ForecastBillableEventRows

return an object wrapping a sql.Rows to allow streaming the results
in the handler. But this approach requires to call tx.Rollback()
once the read is done to avoid leak transactions.

This is obscure and we think it closes the transaction in the wrong
place.

We cannot stop using transactions because `ForecastBillableEventRows`
is designed to inject some fake rows in the `events` table.

Instead of refactoring all the `ForecastBillableEventRows` function,
we ask a context when calling this methods, which will be associated
to the transaction. The caller is responsible to cancel the context
that will trigger the `tx.Rollback()`