Split the API server and event collector
We need to make the API server blue-green-deployable. This means that it
needs to be possible to run it twice, simultaneously, against the same
CF.

This is a problem within the current architecture as the collector and
the API both run in the same process. We can't allow the collector to
run more than once as we're unsure if this might have an effect on the
bills produced.

The net result of this is that we need to split this application into 2,
the API and the collector, so that we can blue-green-deploy the API
portion whilst leaving the deployment of the collector as a simple `cf
push` (which results in no more than one instance running).

We achieve this here by adding a second-level noun to the program's
invocation and starting only those sections of the program as are
required.

To use this: where previously you ran `./path/to/paas-billing`, now to
start the collector you should run `./path/to/paas-billing collector`
and to start the api/event server you should run `./path/to/paas-billing
api`.

It is not possible to run both components at once.