Create custom Postgres types only once
Originally we dropped and recreated every types in the init. Because of this we
had to drop all tables first which used these types, which we want to prevent.

We will create the types only once. If we ever need to alter these types in the
future we should really introduce a proper migration mechanism.