Add quota_definition_guid to the billing event consolidation
The paas-admin cost report currently queries the cf api for the quota guid and the
billing endpoint for the billing events.

The quota guid isn't associated with the billing events so the quota in the
report is unrelated to the values. This updates the consolidation to include
the org's quota_definition_guid so the correct quota is avaiable via the
billing endpoint.