Add database migrations for quota_defintion_guid
We cannot only change the schema in the CREATE TABLE because in prod the
tables already exist.

Instead add a very basic "migration" to add the column to existing
table if it does not exist.

Also remove the not-null constraint because the existing table will have
null entries when we create this column.

Co-authored-by: Mohamed <mohamed.deerow@digital.cabinet-office.gov.uk>
Co-authored-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>