Rename eventserver package to apiserver
This commit renames the `eventserver` package to be named `apiserver`.
This package's only purpose is to handle API requests.

The old name did not stand out next to similarly-named packages such
as `eventio` and `eventstore`. This new name should make it easier to
find relevant code and understand the structure of the project.