Update Makefile to use Go code from `vendor`
Go 1.11 introduced `go mod`. We'd like to switch to using it for fun,
but in particular so that when running tests in Concourse we don't
have to put Go code inside a `GOPATH`. The previous commit switched to
using Go modules but didn't actually make Go use it.

There is a `-mod=vendor` command line argument that can be used to
tell Go 1.11+ to use packages from the `vendor` directory [1].

Ginkgo v1.8.0 added support for passing through the `-mod=vendor`
argument, but I'm not the first to discover that it doesn't seem to
work [2]. To workaround that this commit uses the `GOFLAGS`
environment variable instead, and so doesn't need to upgrade Ginkgo.

[1] https://github.com/golang/go/wiki/Modules#how-do-i-use-vendoring-with-modules-is-vendoring-going-away
[2] https://github.com/onsi/ginkgo/issues/559