Make Refresh Schedule and Timeouts massive
At the moment paas-billing does an "init" when it starts up and a
"refresh" every 30 minutes. In theory therefore, you should be able to
see an application you've just pushed in your bills within the next hour
or so.

Sadly, the "refresh" takes > 25 minutes in London currently, so the
bills are currently empty because it never succeeds.

By running the refresh every 6 hours instead of every 30 minutes, and
allowing the queries to run for 2 hours instead of 25 minutes it should
have plenty of time to complete. The bills should be correct again,
which is the main thing.

This is all on the assumption that having < 1 hour freshness on bills is
not an essential feature. Having < 1 day freshness should be good
enough.