Configure API logging
Echo's main logging only seems to record errors (e.g., HTTP 404.) To
log every request, one needs to use the Logger Middleware. But these
two different things take different types: Echo's Logger is an
`echo.Logger` whereas the Logger Middleware is an `io.Writer`.

We already have a type wrapping `lager.Logger` to implement
`echo.Logger`. This commit makes that type also implement `io.Writer`
so that it can also be used by the Echo Logger Middleware.

An alternative choice would have been to modify the Logger Middleware.
I chose not to do this because we'd have had to maintain that, it
would not remove the need for our existing wrapper around
`lager.Logger`, and it looked relatively complex.

For context, wrapping `lager.Logger` is important so that we get the
same tags in Kibana for both API and Collector log entries.