Allow exchange rate changes at midnight on any day
In commit 6b15db51 we tried to change the USD–to-GBP exchange rate on
October 18th. That change failed with an error:

  new row for relation "currency_rates" violates check constraint
  "valid_from_start_of_month"

As we have discovered, the `pricing_plans`, `vat_rates` and
`currency_rates` tables have all had a constraint requiring changes to
take effect from midnight on the first day of the month. As far as I
know there isn't a technical reason this is needed; it is merely a
convention.

This commit migrates us to a more relaxed constraint. We still
constrain exchange rate changes to take effect at midnight, but we no
longer require them to happen on the first day of a month. This allows
us to move faster.

We still forbidding changes partway through a day. I anticipate that
will be enough to avoid problems when improving how we compute bills.

`paas-admin` will need to be updated so that it can report multiple
exchange rates within one month [1].

[1] https://github.com/alphagov/paas-admin/blob/f89414f3/src/components/statements/statements.njk#L105