Force service_plan.unique ID to be a UUID
The create_events.sql query assumes the service plan unique_id field to
be a UUID / casts it to uuid without checks.
The unique_id is named so because it need not be a UUID, though.

We have encountered a case of a tenant using a custom broker (Vault
broker) which will set unique_id of the form '<uuid>.shared'.
The presence of such service_plan.unique_id breaks the generation of the
events table in paas-billing.

This commit introduces a function uuid_or_placeholder which replaces the
unique_id field with a placeholder uuid 'd5091c33-2f9d-4b15-82dc-4ad69717fc03'
which we also use in case of the unique_id being NULL.

This is feasible as we are not currently using Vault broker for any
billed services and do not care about these custom plans.
Also, work to replace this way of processing events is under way.

pair: @poveyd @schmie