Set a default value for `owner` in `orgs` table
Not all orgs in our production CF database seem to have the owner
metadata set (e.g. tests). This breaks the (re-)generation of the orgs
table in the `billing-db`.
Setting a default `Owner not set` value of `owner` in the `orgs` table
so that the generation of the table does not fail.