Implement flag for directing output
In dev and staging environments, we're not bothered about writing the data to
Google Sheets, but we'd still like to be able to see it. For that reason, we
write to stdout.

In production environments, we want the data sending to Google Sheets, so we
introduce a flag to toggle it.