Add a semi-colon
Make sure each update statement is followed by a semi colon for an
easier readability and an attempt to prevent issues.

This is done the "hard" way, where we're simply joining each statement,
with a semi and a new line purely as there is no obvious way to modify
knex's behaviour.