DROP new billing tables prior to CREATE
The new billing tables contain primary key constraints that cause
collisions/loss of idempotency when re-initialising the database.
Since we consider the billing configuration in paas-cf the source of
truth for these data, the tables are safe to truncate/drop. This
ffollows the DROP TABLE IF EXISTS, CREATE TABLE pattern used in old
billing and suffices for an MVP. We may want to change this is future to
make it more robust/elegant.