Changing the collector to call update_resources
We currently don't have a way to call update_resources which is needed
for the new billing.

This adds this functionality to collector so that it is run on the same
frequency.

It starts from 2016 to ensure that we get all the updates, but we will
probably want to change this later.