Add service_label to plan_name in update_resources
The paas-admin output expects the service label to be prepended to the
plan name in order to make plan names more readable, e.g. `postgres
small-11` rather than `small-11`. We update the plan name as the field
is not used for comparisons.

Also update db_tests of update resources to expect the concatenation