Stop looking for pricing plan components from the beginning of time
We now have components that start after the beginning of time EPOCH.

This however confuses the generateMissingPlans function as it still
looks for things that start at EPOCH. So it the tries to
generate new pricing components, with epoch as the starting time and
fails as it does not match the pricing plan (the logic we fixed in this
commit https://github.com/alphagov/paas-billing/pull/133/commits).

This fixes the logic in the same way to make sure we generate a pricing plan
component only if we need one for the current event, that is there is
not a component valid before the start of the event.