cf billing reconcilliation tool
This tool is to verify that cloud foundries view of the world is a lot
like billings. This is only done for services initially.

It looks at cloud foundry from 4 hours ago, so may miss something, but
this is to take into consideration the time taken for billing to update.

Once it has got a list of services from each API it makes sets of them
using maps.

It then checks the intersection of the maps and the makes sure there are
not too many things in billing and not in cf (which happens for
transient services)