eventstore: use custom aggregator last_agg instead of array manipulation in create_events
when this query is looking to fill null values with those
from prior events, rather than building (sometimes very large)
arrays with array_agg, scanning them for nulls and pulling
one final value out, use a (trivial) custom aggregator that
only ever has to hold a single value in memory.

instead of building the null-detection into the aggregation
function, provide the most basic primitive that will allow us
to do what we need when combined with the FILTER aggregation
clause - performance is approximately the same and this
approach is more flexible (and understandable?).

reversing the direction of the window and keeping track of the
*last* seen non-null value, rather than looking *forward* from
each row in search of the first non-null value encountered
allows the query to run much faster, as the aggregation doesn't
need to be started again for each row. generally, window
functions with an UNBOUNDED PRECEDING frame can be done with a
single aggregation per partition.

reversing the direction of the window also has the advantage
that the plan doesn't need to re-sort everything for the next
window used by event_ranges.