eventstore: reduce candidate valid_* rows in create_events outer joins
many of the rows in the service_plans, services, orgs
& spaces tables don't update any values this query cares
about, therefore these rows can be safely pruned and
valid_for's window function will happily create a single
long range where it would otherwise have created
multiple contiguous ranges that all presented the same
values (for any columns we're interested in).

this gives the outer joins a lot less work to do
(currently it's a hash join with the range-checking done
as a filter, we give it far fewer ranges to check)

to do this however, we need to add a (fairly) simple
custom aggregation function that allows us to compare
neighbouring values for distinctness (i.e. equality
with null-handling) from each other. this generalises
nicely to an "anydistinct" aggregator that has easy to
understand semantics and also behaves sanely when used
with more than two values (provided your type is sanely
defined).

this is necessary because SELECT DISTINCT and SELECT
DISTINCT ON can't *quite* be tricked into doing what
we want here (sort by an unrelated field, have
repetitions if the resource does indeed flip flop
between values).