EventStore.Init(): don't regenerateEvents()
this does mean that some tables aren't guaranteed to be
in place when Init() returns, but these are
consolidation-related and aren't needed by the (purely
data-fetching) other workers the collector process goes
on to start following this. instead, regenerateEvents()
is expected to be run by the event processor loop as
soon as it is started.