App.Init(): don't `store.ConsolidateAll()` before returning
we've removed the initial store `Refresh()` that this call
implicitly relied upon and this now fails given a clean
database to start on.

as with the removal of the initial `Refresh()`, the
`ConsolidateAll()` gets called anyway as part of the event
processing loop and, from an external process/client's point of view,
will still get called just about as soon as it would have otherwise
due to the initial run the event processing loop now does
immediately when started. the only difference is we're starting the
fetching loops before we move on to that.