BillableEventsHandler: don't use multiple db sessions
the attempt to save app memory by splitting the query across
multiple database sessions is done at the expense of database
memory and connections, which is fundamentally a more expensive
resource - currently, requests for big time ranges can easily
exhaust the database's available connections (even if there
isn't much data in the database).

remove the use of multiple connections, but keep the month-
splitting code as it allows us to take advantage of consolidated
blocks (which are significantly faster to retrieve).

fundamentally we need to either put result count limits in
(perhaps a slightly awkward thing to introduce into an existing
API in a backwards-compatible fashion) or perhaps investigate
enforcing use of compressed connections (the use of "streaming"
responses is also a bit tricky because it breaks proper error
handling)